Add a horrible hack to fix movie playback failure
Avoid making the window size an exact match to the screen. This can
be necessary to play fmv sequences via the enhanced video renderer to
a window also used for OpenGL. It seems some graphics drivers special
case the exact match, and use a special mode for OpenGL rendering in
that case, and that prevents the window being used for other types of
output.

Hopefully a better way to make this work will be found at some stage.