Correctly account for alpha in overridden textures
Some of the texture packs use translucency. This commit adds a rough map
of the transparency in such textures so that primitives that make use of
it can be recognised. Such primitives are delayed so that they can be
rendered later over the top of opaque ones with blending. Strictly
speaking the delayed primitives should be sorted back to front, but we
still have the shader drop fragments with 0 alpha so few artifacts occur.